requirements.txt
PyQt6
pyobjc
Pillow